Change the type of theDelay as bytes can only go to 255, and the user may want a longer delay time.

+++ LoopingBoxes/LoopingBoxes_NoDelay/LoopingBoxes_NoDelay.ino
@@ -14,7 +14,7 @@
    User editable variables
 */
 // theDelay tracks how long to pause between colour changes
-byte theDelay = 150;
+int theDelay = 150;
 
 // innerColour sets the colour that is used for the inner square
 rgb_t innerColour = GREEN;
